+++ CHANGELOG.md
@@ -12,6 +12,7 @@ New features:
 - Overriding samplerate also scales chunksize.
 - Use updated faster resampler.
 - Enable experimental neon support in resampler via `neon` feature.
+- Add `Loudness` volume control filter.
 
 
 ## 0.4.0